Mavenize the project into two sub modules for their respective jars
