--- build.xml
@@ -1,55 +1,0 @@
-<project name="ScalyrLogback" default="jar" basedir=".">
-  <!-- global properties for this script -->
-  <property name="src"  location="src"/>
-  <property name="build" location="ant_temp"/>
-  <property name="dist"  location="ant_dist"/>
-
-  <path id="build.classpath">
-    <fileset dir="${src}/../lib/">
-      <include name="*.jar" />
-    </fileset>
-  </path>
-
-  <!-- Initialize the build environment; for use by other targets -->
-  <target name="init">
-    <!-- Create the build timestamp -->
-    <tstamp/>
-
-    <!-- Create our output directories -->
-    <mkdir dir="${build}"/>
-    <mkdir dir="${dist}"/>
-  </target>
-
-  <!-- Compile the java code from ${src} into ${build} -->
-  <target name="compile" depends="init" description="Compile all Java files">
-    <javac srcdir="${src}" destdir="${build}"
-      includeAntRuntime="false"
-      classpathref="build.classpath"
-      debug="true"
-      source="1.6"
-      target="1.6"
-    >
-      <src path="${src}" />
-      <exclude name="**/test/*"/>
-    </javac>
-  </target>
-
-  <!-- Build the API library -->
-  <target name="jar" depends="compile" description="Package the Scalyr logback and log4J Appenders into .jar files.">
-    <jar destfile="${dist}/scalyrLogback.jar">
-       <fileset dir="${build}" includes="com/scalyr/logback/*.class"/>
-       <fileset dir="${src}" includes="com/scalyr/logback/*.java"/>
-   </jar>
-      <jar destfile="${dist}/scalyrLog4J.jar">
-          <fileset dir="${build}" includes="com/scalyr/log4j/*.class"/>
-          <fileset dir="${src}" includes="com/scalyr/log4j/*.java"/>
-      </jar>
-  </target>
-
-  <!-- cleanup target: deletes all non-source files -->
-  <target name="clean" description="delete all generated files" >
-    <delete dir="${build}"/>
-    <delete dir="${dist}"/>
-  </target>
-
-</project>
