+++ scalyr-log4j/src/test/resources/log4j.properties
