+++ scalyr-logback/src/test/resources/logback.groovy
@@ -3,11 +3,11 @@ import ch.qos.logback.core.ConsoleAppender
 import com.scalyr.logback.ScalyrAppender
 
 appender("CONSOLE", ConsoleAppender) {
-  encoder(PatternLayoutEncoder) {
-    pattern = "%-4relative [%thread] - %msg%n"
-  }
+    encoder(PatternLayoutEncoder) {
+        pattern = "%-4relative [%thread] - %msg%n"
+    }
 }
 appender("SCALYR", ScalyrAppender) {
-	apiKey = "YOUR_API_KEY_HERE"
+    apiKey = "YOUR_API_KEY_HERE"
 }
 root(WARN, ["CONSOLE", "SCALYR"])
